refactor: section 패키지 생성 및 SectionLinks의 접근 제어자 수정
