+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 @Repository
 public class SectionDao {
