+++ src/main/java/wooteco/subway/domain/section/SectionLinks.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import java.util.HashMap;
 import java.util.List;
@@ -12,7 +12,7 @@ public class SectionLinks {
         this.sections = new HashMap<>(sections);
     }
 
-    public static SectionLinks from(List<Section> sections) {
+    static SectionLinks from(List<Section> sections) {
         Map<Long, Long> stationIds = new HashMap<>();
         for (Section existingSection : sections) {
             stationIds.put(existingSection.getUpStationId(), existingSection.getDownStationId());
@@ -20,33 +20,33 @@ public class SectionLinks {
         return new SectionLinks(stationIds);
     }
 
-    public boolean isNotExistMatchedStation(Section section) {
+    boolean isNotExistMatchedStation(Section section) {
         return isNotExistStation(section.getUpStationId()) && isNotExistStation(section.getDownStationId());
     }
 
-    public boolean isAllMatchedStation(Section section) {
+    boolean isAllMatchedStation(Section section) {
         return !isNotExistStation(section.getUpStationId()) && !isNotExistStation(section.getDownStationId());
     }
 
-    public boolean isEndSection(Section section) {
+    boolean isEndSection(Section section) {
         return (isExistUpStation(section.getDownStationId()) && !isExistDownStation(section.getDownStationId()))
             || (isExistDownStation(section.getUpStationId()) && !isExistUpStation(section.getUpStationId()));
     }
 
-    public boolean isEndStation(Long stationId) {
+    boolean isEndStation(Long stationId) {
         return (isExistDownStation(stationId) && !isExistUpStation(stationId))
             || (!isExistDownStation(stationId) && isExistUpStation(stationId));
     }
 
-    public boolean isNotExistStation(Long id) {
+    boolean isNotExistStation(Long id) {
         return !isExistUpStation(id) && !isExistDownStation(id);
     }
 
-    public boolean isExistUpStation(Long id) {
+    boolean isExistUpStation(Long id) {
         return sections.containsKey(id);
     }
 
-    public boolean isExistDownStation(Long id) {
+    boolean isExistDownStation(Long id) {
         return sections.containsValue(id);
     }
 }
