+++ src/test/java/wooteco/subway/service/SectionServiceTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import wooteco.subway.dao.SectionDao;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 
 @JdbcTest
 public class SectionServiceTest {
